# optimize: made ignoredHeaderNames map.