--- a/glide.lock
+++ b/glide.lock
@@ -1,40 +1,34 @@
-hash: 8776b72b857522095fe1b5a7f38ac157c6cc71f8cd3d410ce6cf05bbe46c1a64
-updated: 2017-01-26T20:24:29.7292639+09:00
+hash: 9f27eda7c264bb4031d57a6300ad0d3fead43308b537e1be96f23e94b36be412
+updated: 2017-06-14T12:41:19.942947125-07:00
 imports:
 - name: github.com/fukata/golang-stats-api-handler
   version: ab9f90f16caab828afda479fd34bfbbbba2efcee
-  subpackages:
-  - github.com/fukata/golang-stats-api-handler
 - name: github.com/jessevdk/go-flags
   version: 4e64e4a4e2552194cf594243e23aa9baf3b4297e
-  subpackages:
-  - github.com/jessevdk/go-flags
 - name: github.com/lestrrat/go-apache-logformat
   version: aef9f5955f335dc2fbb969c95b849347b15e36c9
-  subpackages:
-  - github.com/lestrrat/go-apache-logformat
 - name: github.com/lestrrat/go-file-rotatelogs
   version: 505c036604b3326a96097ca91865df30043093a5
-  subpackages:
-  - github.com/lestrrat/go-file-rotatelogs
 - name: github.com/lestrrat/go-server-starter-listener
   version: 00dd68592c85334ce94a28199ec7961352c0ba6d
-  subpackages:
-  - github.com/lestrrat/go-server-starter-listener
 - name: github.com/lestrrat/go-strftime
   version: 04ef93e285313c8978cbc7cad26d2aa7a9927451
-  subpackages:
-  - github.com/lestrrat/go-strftime
 - name: github.com/pkg/errors
   version: 248dadf4e9068a0b3e79f02ed0a610d935de5302
-  subpackages:
-  - github.com/pkg/errors
 - name: github.com/renstrom/shortuuid
   version: d728e00b727de969356736fabe7c6f405347a26c
-  subpackages:
-  - github.com/renstrom/shortuuid
 - name: github.com/satori/go.uuid
   version: b061729afc07e77a8aa4fad0a2fd840958f1942a
+testImports:
+- name: github.com/davecgh/go-spew
+  version: 04cdfd42973bb9c8589fd6a731800cf222fde1a9
   subpackages:
-  - github.com/satori/go.uuid
-devImports: []
+  - spew
+- name: github.com/pmezard/go-difflib
+  version: d8ed2627bdf02c080bf22230dbb337003b7aba2d
+  subpackages:
+  - difflib
+- name: github.com/stretchr/testify
+  version: f6abca593680b2315d2075e0f5e2a9751e3f431a
+  subpackages:
+  - assert