--- a/glide.yaml
+++ b/glide.yaml
@@ -4,3 +4,4 @@ import:
 - package: github.com/lestrrat/go-apache-logformat
 - package: github.com/lestrrat/go-server-starter-listener
 - package: github.com/renstrom/shortuuid
+- package: github.com/stretchr/testify