--- a/glide.yaml
+++ b/glide.yaml
@@ -4,4 +4,5 @@ import:
 - package: github.com/lestrrat/go-apache-logformat
 - package: github.com/lestrrat/go-server-starter-listener
 - package: github.com/renstrom/shortuuid
+testImport:
 - package: github.com/stretchr/testify