And added the test and benchmark for **proxy.copyRequest()**.

## Before

```
$ go test -benchmem -benchtime 5s -bench .
BenchmarkCopyRequest-8           2000000              3376 ns/op            1655 B/op         15 allocs/op
PASS
ok      github.com/kazeburo/chocon/proxy        10.254s
```

## After

```
$ go test -benchmem -benchtime 5s -bench .
BenchmarkCopyRequest-8          10000000               970 ns/op             816 B/op          6 allocs/op
PASS
ok      github.com/kazeburo/chocon/proxy        10.710s
```